Headful mode seems to also have problems verifying tags after the grace period. Moved the headful tests back to headless class and attempted to select the items programmatically instead.
