+++ src/test/java/guitests/guihandles/GuiHandle.java
@@ -90,7 +90,7 @@ public class GuiHandle {
     }
 
     public void pressEnter() {
-        guiRobot.type(KeyCode.ENTER).sleep(1000);
+        guiRobot.type(KeyCode.ENTER).sleep(500);
     }
 
     protected void pressEsc() {
